requirements.txt
accelerate==1.11.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
ascii_colors==0.11.4
async-timeout==5.0.1
attrs==25.4.0
av==16.0.1
beartype==0.22.5
beautifulsoup4==4.14.2
bitsandbytes==0.48.2
boto3==1.40.69
botocore==1.40.69
brotli==1.2.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
coloredlogs==15.0.1
colorlog==6.10.1
CoLT5-attention==0.11.1
configparser==7.2.0
contourpy==1.3.2
cryptography==46.0.3
cssselect2==0.8.0
cycler==0.12.1
dill==0.4.0
distro==1.9.0
doclayout_yolo==0.0.4
einops==0.8.1
einops-exts==0.0.4
einx==0.3.0
exceptiongroup==1.3.0
fast-langdetect==0.2.5
fastapi==0.121.1
fasttext-predict==0.9.2.4
ffmpy==0.6.4
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
frozendict==2.4.6
frozenlist==1.8.0
fsspec==2025.10.0
ftfy==6.3.1
future==1.0.0
google-auth==2.43.0
google-genai==1.49.0
gradio==5.49.1
gradio_client==1.13.3
gradio_pdf==0.0.22
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-retries==0.4.5
huggingface-hub==0.36.0
humanfriendly==10.0
hyper-connections==0.2.1
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
jiter==0.11.1
jmespath==1.0.1
joblib==1.3.2
json_repair==0.53.0
kiwisolver==1.4.9
lazy_loader==0.4
lightrag-hku==1.4.9.8
local-attention==1.11.2
loguru==0.7.3
magika==0.6.3
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mineru==2.6.4
mineru_vl_utils==0.1.15
modelscope==1.31.0
mpmath==1.3.0
multidict==6.7.0
nano-vectordb==0.0.4.3
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
ollama==0.6.0
omegaconf==2.3.0
onnxruntime==1.23.2
openai==2.7.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pdfminer.six==20250506
pdftext==0.6.3
pillow==11.3.0
pipmaster==1.0.10
polars==1.35.2
polars-runtime-32==1.35.2
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycparser==2.23
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydub==0.25.1
pydyf==0.11.0
Pygments==2.19.2
pyparsing==3.2.5
pypdf==6.2.0
pypdfium2==4.30.0
pyphen==0.17.2
pypinyin==0.55.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
qwen==0.1.1
qwen-vl-utils==0.0.14
regex==2025.11.3
reportlab==4.4.4
requests==2.32.5
rich==14.2.0
robust-downloader==0.0.2
rsa==4.9.1
ruff==0.14.4
s3transfer==0.14.0
safehttpx==0.1.7
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.5.2
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==5.1.2
shapely==2.1.2
shellingham==1.5.4
simsimd==6.5.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.49.3
stringzilla==4.2.3
sympy==1.14.0
tenacity==9.1.2
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
tifffile==2025.5.10
tiktoken==0.12.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.22.1
tomlkit==0.13.3
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.57.1
triton==3.5.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.226
ultralytics-thop==2.0.18
urllib3==2.5.0
uvicorn==0.38.0
vector-quantize-pytorch==1.25.1
wcwidth==0.2.14
weasyprint==66.0
webencodings==0.5.1
websockets==15.0.1
xlsxwriter==3.2.9
yarl==1.22.0
zetascale==2.8.8
zopfli==0.4.0
